+++ src/apply.ts
@@ -1,21 +1,22 @@
 import * as CachePatch from "./cache-patch";
 import * as GraphQLEntityCache from "graphql-norm";
+import * as GraphQLNormStale from "graphql-norm-stale";
 
 interface MutableStaleEntities {
   // tslint:disable-next-line:readonly-keyword
-  [key: string]: GraphQLEntityCache.StaleEntity | undefined;
+  [key: string]: GraphQLNormStale.StaleFields | undefined;
 }
 
 interface MutableEntityCache {
   // tslint:disable-next-line:readonly-keyword
-  [id: string]: GraphQLEntityCache.Entity;
+  [id: string]: GraphQLEntityCache.NormObj;
 }
 
 export function apply(
   patches: ReadonlyArray<CachePatch.CachePatch>,
-  cache: GraphQLEntityCache.EntityCache,
-  staleEntities: GraphQLEntityCache.StaleEntities
-): [GraphQLEntityCache.EntityCache, GraphQLEntityCache.StaleEntities] {
+  cache: GraphQLEntityCache.NormMap,
+  staleEntities: GraphQLNormStale.StaleMap
+): [GraphQLEntityCache.NormMap, GraphQLNormStale.StaleMap] {
   if (patches.length === 0) {
     return [cache, staleEntities];
   }
@@ -73,12 +74,12 @@ export function apply(
 
 function applyInvalidateEntity(
   patch: CachePatch.InvalidateEntity,
-  cache: GraphQLEntityCache.EntityCache,
+  cache: GraphQLEntityCache.NormMap,
   staleEntities: MutableStaleEntities
 ): void {
   const entity = cache[patch.id];
   if (entity !== undefined) {
-    const newStaleEntity: Mutable<GraphQLEntityCache.StaleEntity> = {
+    const newStaleEntity: Mutable<GraphQLNormStale.StaleFields> = {
       ...staleEntities[patch.id]
     };
     for (const entityKey of Object.keys(entity)) {
@@ -93,7 +94,7 @@ function applyInvalidateEntity(
 
 function applyInvalidateField(
   patch: CachePatch.InvalidateField,
-  cache: GraphQLEntityCache.EntityCache,
+  cache: GraphQLEntityCache.NormMap,
   staleEntities: MutableStaleEntities
 ): void {
   if (cache[patch.id] !== undefined) {
@@ -125,9 +126,9 @@ function applyInvalidateField(
 export declare type Mutable<T> = { -readonly [P in keyof T]: T[P] };
 
 function invalidateRecursive(
-  cache: GraphQLEntityCache.EntityCache,
+  cache: GraphQLEntityCache.NormMap,
   staleEntities: MutableStaleEntities,
-  startingEntity: GraphQLEntityCache.EntityFieldValue | null
+  startingEntity: GraphQLEntityCache.NormFieldValue | null
 ): void {
   if (
     typeof startingEntity === "number" ||
@@ -153,7 +154,7 @@ function invalidateRecursive(
       continue;
     }
     const entityFieldKeys = Object.keys(entity);
-    const newStaleEntity: Mutable<GraphQLEntityCache.StaleEntity> = {
+    const newStaleEntity: Mutable<GraphQLNormStale.StaleFields> = {
       ...staleEntities[entityId]
     };
 
@@ -176,9 +177,9 @@ function invalidateRecursive(
 }
 
 function isArrayOfEntityIds(
-  cache: GraphQLEntityCache.EntityCache,
-  field: GraphQLEntityCache.EntityFieldValue | null
-): field is ReadonlyArray<GraphQLEntityCache.EntityId> {
+  cache: GraphQLEntityCache.NormMap,
+  field: GraphQLEntityCache.NormFieldValue | null
+): field is ReadonlyArray<GraphQLEntityCache.NormKey> {
   if (Array.isArray(field) && field.some(x => !!cache[x])) {
     return true;
   }
@@ -272,7 +273,7 @@ function applyRemoveEntityElement(
   if (entityAndFieldExists(cache, patch)) {
     const fieldNameWithArgs = withArgs(patch.fieldName, patch.fieldArguments);
     const arr = cache[patch.id][fieldNameWithArgs] as ReadonlyArray<
-      GraphQLEntityCache.EntityId
+      GraphQLEntityCache.NormKey
     >;
     // Shallow mutation of cache OK as we have a shallow copy
     cache[patch.id] = {
@@ -283,7 +284,7 @@ function applyRemoveEntityElement(
 }
 
 function entityAndFieldExists(
-  cache: GraphQLEntityCache.EntityCache,
+  cache: GraphQLEntityCache.NormMap,
   patch: {
     readonly id: string;
     readonly fieldName: string;
@@ -298,7 +299,7 @@ function entityAndFieldExists(
 }
 
 function entityExists(
-  cache: GraphQLEntityCache.EntityCache,
+  cache: GraphQLEntityCache.NormMap,
   patch: { readonly id: string }
 ): boolean {
   return !!cache[patch.id];