+++ CHANGELOG.md
@@ -5,7 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [Unreleased](https://github.com/dividab/graphql-norm-patch/compare/v0.14.0...master)
+## [Unreleased](https://github.com/dividab/graphql-norm-patch/compare/v0.16.0...master)
+
+## [v0.16.0](https://github.com/dividab/graphql-norm-patch/compare/v0.15.0...v0.16.0) - 2019-09-29
 
 ### Changed